Switch to a new way of setting the resolver on the validation factory
This is necessary, because since Laravel 5.2 the validator service provider is
deferred. We can't set it once at the beginning. Now it gets set in a callback
whenever someone resolves `validator` from the service container.

This also solves #18. We don't really need to defer anymore, because if the
validator service is not used, this service provider essentially only adds a
single entry to an array in the service container.